scripts/invoice-parser/requirements.txt
# Invoice Parser Requirements
# Python 3.8+ required

# PDF Processing
pdfplumber==0.11.5
PyPDF2==3.0.1
pdf2image==1.17.0

# OCR Support
pytesseract==0.3.13
pillow==11.1.0

# Excel Processing
xlrd==2.0.1
openpyxl==3.1.5

# Document Processing
python-docx==1.2.0

# Data Processing
pandas==2.2.3
numpy==2.2.4

# System Dependencies (install via apt)
# tesseract-ocr
# poppler-utils
# libreoffice (for .doc conversion)